requirements.txt
#git+https://github.com/Rapptz/discord.py@rewrite
Flask
Flask-SQLAlchemy
requests
discord.py==0.16.12